[S07P22B301-307] Fix. [검증]: [회원 관리 기능 테스트]

See merge request s07-ai-image-sub2/S07P22B301!205

+++ back/src/main/java/com/gdcd/back/service/user/UserServiceImpl.java
@@ -107,8 +107,8 @@ public class UserServiceImpl implements UserService {
             }
             user.update(PROFILE_REQUEST_URI, filePath, nickname);
 
-            modifyFollower(target, user);
-            modifyFollowing(target, user);
+            modifyFollower(target.simplify(), user.simplify());
+            modifyFollowing(target.simplify(), user.simplify());
             modifyBlocker(target, user);
             modifyBlocking(target, user);
             modifyReporter(target, user);
@@ -220,14 +220,30 @@ public class UserServiceImpl implements UserService {
     public Map<String, Object> followUser(String token, Long userId) {
         RESULT_OBJECT = new HashMap<>();
         try {
-            User follower = findUserByEmail(decodeToken(token));
-            User following = findUserById(userId);
+            UserSimple follower = findUserByEmail(decodeToken(token)).simplify();
+            UserSimple following = findUserById(userId).simplify();
+
+            if (follower.getId().equals(userId))
+                throw new Exception("SELF-FOLLOW NOT ALLOWED");
+
             Follow follow;
             if (followRepository.findByFollowerAndFollowing(follower, following).isPresent()) {
+                User followerUser = findUserByEmail(decodeToken(token));
+                followerUser.subFollowingCount();
+                userRepository.save(followerUser);
+                User followingUser = findUserById(userId);
+                followingUser.subFollowerCount();
+                userRepository.save(followingUser);
                 followRepository.deleteByFollowerAndFollowing(follower, following);
                 RESULT_OBJECT.put("unfollow", userId);
                 // fix) unFollow할 때는, user의 followCount--
             } else {
+                User followerUser = findUserByEmail(decodeToken(token));
+                followerUser.addFollowingCount();
+                userRepository.save(followerUser);
+                User followingUser = findUserById(userId);
+                followingUser.addFollowerCount();
+                userRepository.save(followingUser);
                 follow = Follow.builder()
                         .follower(follower)
                         .following(following)
@@ -251,7 +267,7 @@ public class UserServiceImpl implements UserService {
                 following = findUserByEmail(decodeToken(token));
             else
                 following = findUserById(userId);
-            List<Follow> documentList = followRepository.findAllByFollowing(following);
+            List<Follow> documentList = followRepository.findAllByFollowing(following.simplify());
             List<FollowListResponseDto> list = new ArrayList<>();
             for (Follow follow : documentList) {
                 list.add(new FollowListResponseDto(follow.getFollower()));
@@ -274,7 +290,7 @@ public class UserServiceImpl implements UserService {
                 follower = findUserByEmail(decodeToken(token));
             else
                 follower = findUserById(userId);
-            List<Follow> documentList = followRepository.findAllByFollower(follower);
+            List<Follow> documentList = followRepository.findAllByFollower(follower.simplify());
             List<FollowListResponseDto> list = new ArrayList<>();
             for (Follow follow : documentList) {
                 list.add(new FollowListResponseDto(follow.getFollowing()));
@@ -312,14 +328,14 @@ public class UserServiceImpl implements UserService {
             throw new Exception("User Not Found");
     }
 
-    private void modifyFollower(User target, User user) {
+    private void modifyFollower(UserSimple target, UserSimple user) {
         List<Follow> followerList = followRepository.findAllByFollower(target);
         for (Follow follow : followerList) {
             follow.modifyFollower(user);
             followRepository.save(follow);
         }
     }
-    private void modifyFollowing(User target, User user) {
+    private void modifyFollowing(UserSimple target, UserSimple user) {
         List<Follow> followingList = followRepository.findAllByFollowing(target);
         for (Follow follow : followingList) {
             follow.modifyFollowing(user);
